pyproject.toml
[project]
name = "emailreader-app"
version = "0.1.0"
description = "A FastAPI app for reading relevant emails."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "celery[redis]",
    "httpx",
    "boto3",
    "aioboto3",
    "msal",
    "requests",
    "python-dateutil",
    "email-validator",
    "asyncpg",
    "eventlet",
    "python-jose[cryptography]",
    "python-multipart"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "moto[s3]>=4.0.0",
    "testcontainers[postgres]>=3.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
